KS-2-4 changed percentage calculation logic (added to.Double)

+++ src/main/kotlin/Lesson2/lesson2_task4.kt
@@ -5,7 +5,7 @@ const val PERCENT_FORMATTER = 100
 
 fun main() {
 
-    val buffCoefficient = BONUS_PERCENTAGE / PERCENT_FORMATTER
+    val buffCoefficient = BONUS_PERCENTAGE.toDouble() / PERCENT_FORMATTER
     val buffIsOn = true
     val crystalOre = 7
     val ironOre = 11
